server/backend/requirements.txt
asgiref==3.6.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pytz==2022.7
sqlparse==0.2.4
djangorestframework-simplejwt==5.2.2
djangorestframework-jwt==1.11.0
drf-jwt==1.19.2
requests==2.27.1
